Refactoring highlights and put all highlight-related functions to utils
